# Entrega 1/backend/src/main/java/com/universidad/solicitudes/application/dto/AsignacionGestorDTO.java
package com.universidad.solicitudes.application.dto;

import java.time.LocalDateTime;
import java.util.UUID;

public record AsignacionGestorDTO(
        UUID id,
        UUID solicitudId,
        UUID gestorId,
        String gestorNombre,
        LocalDateTime fechaAsignacion,
        boolean activa
) {}