# Entrega 1/backend/src/main/java/com/universidad/solicitudes/application/dto/RegistroSeguimientoCrearDTO.java
package com.universidad.solicitudes.application.dto;

import jakarta.validation.constraints.NotBlank;

public record RegistroSeguimientoCrearDTO(
        @NotBlank(message = "El comentario es obligatorio")
        String comentario
) {}